Reprocessing.Tests.Reprocessing.cs reconstructed
- big Test is separated into small test methods
- test methods splitted into two test classes - for unique constraints
  and for deadlocks
- every test has limit of time to execute now, not more that 2 seconds
  (to prevent test infrastructure hanging)
